Skip delete cache if bank account id is null

+++ backend/src/main/java/com/bulletjournal/repository/TransactionDaoJpa.java
@@ -492,7 +492,9 @@ public class TransactionDaoJpa extends ProjectItemDaoJpa<TransactionContent> {
         BankAccount bankAccount = this.bankAccountDaoJpa.getBankAccount(requester, bankAccountId);
         transaction.setBankAccount(bankAccount);
         transaction = this.transactionRepository.save(transaction);
-        this.bankAccountBalanceRepository.deleteById(bankAccountId);
+        if (bankAccountId != null) {
+            this.bankAccountBalanceRepository.deleteById(bankAccountId);
+        }
         return transaction;
     }
